requirement.txt
dlt==1.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
fastapi==0.115.0
uvicorn==0.30.6